WIP: Add test for construction match count
This commit adds a test for construction matches count and updates
the count_matched_pair function so that it also works for construction
matches. The refactored function may still work for forward and
backward matches, but the tests were failing due to a filepath error
before the refactor so not able to confirm.