321 flag to ignore response (#18)
* Upload test data for link filters

* build: function to filter rows to ignore values from link

The function adds a new column indicating if row should be ignored. It is based on set_index and index comparison.

Added 2 tests, one to check the output, and if exception is raised when columns do not match.

* build: function to filter rows to ignore values from link

The function adds a new column indicating if row should be ignored. It is based on set_index and index comparison.

Added 2 tests, one to check the output, and if exception is raised when columns do not match.